update and remove from model
